# Increase performance delete old metrics query